+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Fix default theme color when no color specified.
 - Added X in Y filter.
 - Added save table dialog.
 - History provider to replace history store.